requirements.txt
python-decouple
Telethon==1.24.0
cryptg
heroku3
sqlalchemy==1.3.20
psycopg2-binary
gitpython
asyncio
aiohttp
